now it compiles

+++ app/src/main/java/site/meowcat/loglet/data/TrackPoint.java
@@ -6,7 +6,8 @@ import androidx.room.PrimaryKey;
 @Entity
 public class TrackPoint {
     @PrimaryKey(autoGenerate = true)
-    public int id;
+    public long id;
+
     public long trackId;
     public double latitude;
     public double longitude;
